requirements.txt
appdirs
click
geographiclib
PILLOW
pycairo
requests
s2sphere
